Fix precision issues

+++ src/services/src/Demoulas.ProfitSharing.Data/Entities/Country.cs
@@ -1,7 +1,7 @@
 ﻿using Demoulas.ProfitSharing.Data.Entities.Base;
 
 namespace Demoulas.ProfitSharing.Data.Entities;
-public sealed class Country : LookupTable<byte>
+public sealed class Country : LookupTable<short>
 {
     public required string ISO { get; init; }
     public required string TelephoneCode { get; init; }
